token 有效期修改

+++ ruoyi-system/src/main/java/com/ruoyi/system/domain/BaoBiao/dto/dev/MddSubItemDTO.java
@@ -3,6 +3,7 @@ package com.ruoyi.system.domain.BaoBiao.dto.dev;
 import lombok.Data;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 
 @Data
 public class MddSubItemDTO implements Serializable {
@@ -11,8 +12,8 @@ public class MddSubItemDTO implements Serializable {
     private String unit_nameJSON;
     private String unit_code;
     private Integer productionData;
-    private Integer expandData;
-    private Integer footageData;
+    private BigDecimal expandData;
+    private BigDecimal footageData;
     private Integer enterWellNum;
     private Integer comeOutWellNum;
 }
