+++ ruoyi-system/src/main/java/com/ruoyi/system/domain/BaoBiao/vo/plan/MinePlanJu.java
@@ -4,6 +4,7 @@ import lombok.Data;
 import org.apache.ibatis.type.Alias;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 
 @Alias("MinePlanJu")  // 起一个项目里独一无二的名字
 @Data
@@ -14,8 +15,8 @@ public class MinePlanJu implements Serializable {
     private String unitName;
     //一班
     private Integer oneProductionData;//生产数据
-    private Integer oneExpandData;//开拓
-    private Integer oneFootageData;//进尺
+    private BigDecimal oneExpandData;//开拓
+    private BigDecimal oneFootageData;//进尺
     private Integer oneEnterWellNum;//入井人数
     private Integer oneComeOutWellNum;//出境人数
     private Integer oneTotalDownCount;//入井人数
@@ -23,16 +24,16 @@ public class MinePlanJu implements Serializable {
 
     //二班
     private Integer twoProductionData;//生产数据
-    private Integer twoExpandData;//开拓
-    private Integer twoFootageData;//进尺
+    private BigDecimal twoExpandData;//开拓
+    private BigDecimal twoFootageData;//进尺
     private Integer twoEnterWellNum;//入井人数
     private Integer twoComeOutWellNum;//出境人数
     private Integer twoTotalDownCount;//入井人数
     private Integer twoTotalUpCount;//出井人数
     //三班
     private Integer threeProductionData;//生产数据
-    private Integer threeExpandData;//开拓
-    private Integer threeFootageData;//进尺
+    private BigDecimal threeExpandData;//开拓
+    private BigDecimal threeFootageData;//进尺
     private Integer threeEnterWellNum;//入井人数
     private Integer threeComeOutWellNum;//出境人数
     private Integer threeTotalDownCount;//入井人数
